requirements.txt
serpapi
pandas
requests
beautifulsoup4
apache-airflow
gspread
oauth2client
scikit-learn
python-dotenv
langchain-core
langchain-openai